remove unused dependencies
