+++ WolvenKit/App.xaml.cs
@@ -6,13 +6,16 @@ using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
 using System.Windows;
+using Catel;
 using Catel.IoC;
 using Catel.Logging;
 using Catel.Messaging;
+using Catel.MVVM;
 using CP77.CR2W;
 using HandyControl.Tools;
 using NodeNetwork;
-using Orchestra.Services;
+using ProtoBuf.Meta;
+using WolvenKit.Common;
 using WolvenKit.Common.Oodle;
 using WolvenKit.Common.Services;
 using WolvenKit.Common.Tools.Oodle;
@@ -21,11 +24,13 @@ using WolvenKit.Functionality.Controllers;
 using WolvenKit.Functionality.Helpers;
 using WolvenKit.Functionality.Initialization;
 using WolvenKit.Functionality.Services;
+using WolvenKit.Functionality.WKitGlobal;
 using WolvenKit.Functionality.WKitGlobal.Helpers;
 using WolvenKit.Modkit.RED3;
 using WolvenKit.Modkit.RED4;
 using WolvenKit.Modkit.RED4.RigFile;
 using WolvenKit.RED4.CR2W;
+using WolvenKit.RED4.CR2W.Archive;
 using WolvenKit.Views;
 using WolvenManager.Installer.Services;
 
@@ -45,16 +50,21 @@ namespace WolvenKit
             // Set application licenses.
             Initializations.InitializeLicenses();
 
+            //protobuf
+            RuntimeTypeModel.Default[typeof(IGameArchive)].AddSubType(20, typeof(Archive));
+
 
             var serviceLocator = ServiceLocator.Default;
 
 
             // Orchestra
-            serviceLocator.RegisterType<IAboutInfoService, AboutInfoService>();
-            serviceLocator.RegisterType<INotificationService, NotificationService>();
 
 
             // Wkit
+
+            serviceLocator.RegisterType<INotificationService, NotificationService>();
+
+
             var config = SettingsManager.Load();
             serviceLocator.RegisterInstance(typeof(ISettingsManager), config);
 
@@ -63,7 +73,6 @@ namespace WolvenKit
             serviceLocator.RegisterType<IUpdateService, UpdateService>();
 
             // singletons
-            serviceLocator.RegisterType<IApplicationInitializationService, ApplicationInitializationService>();
             serviceLocator.RegisterType<IHashService, HashService>();
 
             serviceLocator.RegisterTypeAndInstantiate<IProjectManager, ProjectManager>();
@@ -99,7 +108,6 @@ namespace WolvenKit
             // Set service locator.
             var serviceLocator = ServiceLocator.Default;
 
-            serviceLocator.RegisterType<IRibbonService, RibbonService>();
 
 #if DEBUG
             LogManager.AddDebugListener(false);
@@ -150,6 +158,77 @@ namespace WolvenKit
             // Create WebView Data Folder.
             //Directory.CreateDirectory(@"C:\WebViewData");
             // Message system for video tool.
+        }
+
+        private void InitializeCommands()
+        {
+            var _commandManager = ServiceLocator.Default.ResolveType<ICommandManager>();
+
+            // global commands
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.Exit));
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.About));
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Settings), nameof(AppCommands.Settings.General));
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.Options));
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.BugReport));
+
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.NewProject));
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.PublishProject));
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.OpenProject));
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.OpenLink));
+
+            // application-wide commands that viewmodels can subscribe to
+            // Workspace Viewmodel
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.DelProject));
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.SaveAsProject));
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.SaveProject));
+
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.SaveFile));
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.SaveAll));
+
+
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.ShowAbout));
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.ShowFeedback));
+            _commandManager.CreateCommandWithGesture(typeof(AppCommands.Application), nameof(AppCommands.Application.ShowSettings));
+
+            _commandManager.CreateCommand(AppCommands.Application.ShowLog);
+            _commandManager.CreateCommand(AppCommands.Application.ShowProjectExplorer);
+            _commandManager.CreateCommand(AppCommands.Application.ShowImportUtility);
+            _commandManager.CreateCommand(AppCommands.Application.ShowProperties);
+            _commandManager.CreateCommand(AppCommands.Application.ShowPackageInstaller);
+            _commandManager.CreateCommand(AppCommands.Application.ShowMimicsTool);
+            _commandManager.CreateCommand(AppCommands.Application.ShowCR2WEditor);
+            _commandManager.CreateCommand(AppCommands.Application.ShowImportExportTool);
+
+            _commandManager.CreateCommand(AppCommands.Application.ShowAssetBrowser);
+            _commandManager.CreateCommand(AppCommands.Application.ShowBulkEditor);
+            _commandManager.CreateCommand(AppCommands.Application.ShowCsvEditor);
+            _commandManager.CreateCommand(AppCommands.Application.ShowHexEditor);
+            _commandManager.CreateCommand(AppCommands.Application.ShowJournalEditor);
+            _commandManager.CreateCommand(AppCommands.Application.ShowVisualEditor);
+            _commandManager.CreateCommand(AppCommands.Application.ShowAnimationTool);
+            _commandManager.CreateCommand(AppCommands.Application.ShowAudioTool);
+            _commandManager.CreateCommand(AppCommands.Application.ShowVideoTool);
+
+            _commandManager.CreateCommand(AppCommands.Application.ShowCodeEditor);
+
+            _commandManager.CreateCommand(AppCommands.Application.ShowImporterTool);
+            _commandManager.CreateCommand(AppCommands.Application.ShowCR2WToTextTool);
+            _commandManager.CreateCommand(AppCommands.Application.ShowGameDebuggerTool);
+            _commandManager.CreateCommand(AppCommands.Application.ShowMenuCreatorTool);
+            _commandManager.CreateCommand(AppCommands.Application.ShowPluginManager);
+            //_commandManager.CreateCommand(AppCommands.Application.ShowRadishTool);
+            _commandManager.CreateCommand(AppCommands.Application.ShowWccTool);
+
+            _commandManager.CreateCommand(AppCommands.Application.OpenFile);
+            _commandManager.CreateCommand(AppCommands.Application.NewFile);
+            _commandManager.CreateCommand(AppCommands.Application.PackMod);
+            _commandManager.CreateCommand(AppCommands.Application.BackupMod);
+            _commandManager.CreateCommand(AppCommands.Application.PublishMod);
+
+            _commandManager.CreateCommand(AppCommands.ProjectExplorer.Refresh);
+            _commandManager.CreateCommand(AppCommands.Application.FileSelected);
+
+            _commandManager.CreateCommand(AppCommands.Application.ViewSelected);
         }
     }
 }
