--- WolvenKit/ModuleInitializer.cs
@@ -1,31 +1,0 @@
-using System;
-using System.IO;
-using Catel.IoC;
-using CP77.CR2W;
-using Orchestra.Services;
-using WolvenKit.Bundles;
-using WolvenKit.Functionality.Controllers;
-using WolvenKit.Common.Services;
-using WolvenKit.Common.Tools.Oodle;
-using WolvenKit.Functionality.Services;
-using WolvenKit.Modkit.RED3;
-using WolvenKit.MVVM.Model;
-using WolvenKit.RED4.CR2W;
-using WolvenKit.Modkit.RED4.RigFile;
-using WolvenKit.Core.Services;
-using WolvenKit.Modkit.RED4;
-
-/// <summary>
-/// Used by the ModuleInit. All code inside the Initialize method is ran as soon as the assembly is loaded.
-/// </summary>
-public static class ModuleInitializer
-{
-    /// <summary>
-    /// Initializes the module.
-    /// </summary>
-    public static void Initialize()
-    {
-        
-
-    }
-}
