+++ WolvenKit/Views/Dialogs/MaterialsRepositoryDialog.xaml.cs
@@ -155,14 +155,15 @@ namespace WolvenKit.Views.Dialogs
 
         private void MaterialsButton_Click(object sender, System.Windows.RoutedEventArgs e)
         {
-            CommonOpenFileDialog dialog = new CommonOpenFileDialog { IsFolderPicker = true };
-            if (dialog.ShowDialog() == CommonFileDialogResult.Ok)
-            {
-                MaterialsTextBox.SetCurrentValue(System.Windows.Controls.TextBox.TextProperty, dialog.FileName);
-                MaterialsDepotPath = dialog.FileName;
-                _settingsManager.MaterialRepositoryPath = dialog.FileName;
-                _settingsManager.Save();
-            }
+            //TODO
+            //var dialog = new CommonOpenFileDialog { IsFolderPicker = true };
+            //if (dialog.ShowDialog() == CommonFileDialogResult.Ok)
+            //{
+            //    MaterialsTextBox.SetCurrentValue(System.Windows.Controls.TextBox.TextProperty, dialog.FileName);
+            //    MaterialsDepotPath = dialog.FileName;
+            //    _settingsManager.MaterialRepositoryPath = dialog.FileName;
+            //    _settingsManager.Save();
+            //}
         }
 
         private void ArchivesButton_Click(object sender, System.Windows.RoutedEventArgs e)
